+++ src/Funnels/FunnelStep.php
@@ -10,7 +10,7 @@ use Falcon\Analytics\Models\Event;
 /**
  * One step of a funnel: a weighted milestone matched either by a named event,
  * by a pageview route, or by any of several parallel branches. The value is
- * the step's weight.
+ * the step's weight, **a whole number of points** like every score here.
  *
  * @internal a step is built by `Funnel::step()` and never by hand · the fluent
  *           call hands back the funnel, not the step.
@@ -20,7 +20,7 @@ final readonly class FunnelStep
     /** @param  list<FunnelBranch>  $branches */
     public function __construct(
         public string $label,
-        public float $value,
+        public int $value,
         public ?string $event = null,
         public ?string $route = null,
         public array $branches = [],
